LayoutTests/fast/scrolling/scroll-anchoring-in-overflow-with-page-scale-expected.txt

PASS scroller.scrollTop is 400
PASS successfullyParsed is true

TEST COMPLETE
